Move lhlo_gpu dialect into its own directory.
PiperOrigin-RevId: 407578923
