+++ tensorflow/compiler/mlir/hlo/include/mlir-hlo/Dialect/CMakeLists.txt
@@ -14,4 +14,5 @@
 # limitations under the License.
 #
 add_subdirectory(disc-ral)
+add_subdirectory(lhlo_gpu)
 add_subdirectory(mhlo)
