+++ tensorflow/compiler/mlir/hlo/include/mlir-hlo/Dialect/lhlo_gpu/CMakeLists.txt
@@ -13,6 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-add_subdirectory(disc-ral)
-add_subdirectory(lhlo_gpu)
-add_subdirectory(mhlo)
+add_subdirectory(IR)
