+++ tensorflow/compiler/mlir/hlo/include/mlir-hlo/Dialect/lhlo_gpu/IR/lhlo_gpu_ops.td
@@ -23,9 +23,9 @@ include "mlir/IR/OpBase.td"
 include "mlir/Interfaces/SideEffectInterfaces.td"
 include "mlir-hlo/Dialect/mhlo/IR/hlo_ops_base.td"
 include "mlir-hlo/Dialect/mhlo/IR/lhlo_ops_base.td"
-include "mlir-hlo/Dialect/mhlo/IR/lhlo_gpu_ops_base.td"
-include "mlir-hlo/Dialect/mhlo/IR/lhlo_gpu_ops_enums.td"
-include "mlir-hlo/Dialect/mhlo/IR/lhlo_gpu_ops_structs.td"
+include "mlir-hlo/Dialect/lhlo_gpu/IR/lhlo_gpu_ops_base.td"
+include "mlir-hlo/Dialect/lhlo_gpu/IR/lhlo_gpu_ops_enums.td"
+include "mlir-hlo/Dialect/lhlo_gpu/IR/lhlo_gpu_ops_structs.td"
 
 class LHLOGPU_Op<string mnemonic, list<OpTrait> traits = []> :
   Op<LHLO_GPU_Dialect, mnemonic,
