+++ tensorflow/compiler/mlir/hlo/include/mlir-hlo/Dialect/mhlo/IR/CMakeLists.txt
@@ -37,24 +37,18 @@ mlir_tablegen(hlo_ops_base_attrs.h.inc -gen-attrdef-decls)
 mlir_tablegen(hlo_ops_base_attrs.cc.inc -gen-attrdef-defs)
 add_public_tablegen_target(MLIRhlo_opsIncGen)
 
-function(add_mlir_hlo_dialect_separate_files dialect has_enums)
+function(add_mlir_hlo_dialect_separate_files dialect)
   set(LLVM_TARGET_DEFINITIONS ${dialect}.td)
   mlir_tablegen(${dialect}.h.inc -gen-op-decls)
   mlir_tablegen(${dialect}.cc.inc -gen-op-defs)
   set(LLVM_TARGET_DEFINITIONS ${dialect}_structs.td)
   mlir_tablegen(${dialect}_structs.h.inc -gen-struct-attr-decls)
   mlir_tablegen(${dialect}_structs.cc.inc -gen-struct-attr-defs)
-  if(${has_enums})
-    set(LLVM_TARGET_DEFINITIONS ${dialect}_enums.td)
-    mlir_tablegen(${dialect}_enums.h.inc -gen-enum-decls)
-    mlir_tablegen(${dialect}_enums.cc.inc -gen-enum-defs)
-  endif()
   add_public_tablegen_target(MLIR${dialect}IncGen)
   add_dependencies(mlir-headers MLIR${dialect}IncGen)
 endfunction()
 
-add_mlir_hlo_dialect_separate_files(lhlo_ops NO)
-add_mlir_hlo_dialect_separate_files(lhlo_gpu_ops YES)
+add_mlir_hlo_dialect_separate_files(lhlo_ops)
 
 add_mlir_interface(infer_shape_equality_op_interface)
 add_mlir_interface(lhlo_structured_interface)
