+++ tensorflow/compiler/mlir/hlo/tools/mlir-hlo-opt/CMakeLists.txt
@@ -20,6 +20,7 @@ set(LIBS
         ${conversion_libs}
         MLIROptLib
 
+        LmhloGPUDialect
         MhloRegisterDialects
         AllDiscRalPasses
         AllMhloPasses
